requirements.txt
langgraph
playwright
html2text
pandas
google-generativeai
colorama
python-dotenv
litellm
playwright
agentql
prisma
pydantic
streamlit>=1.28
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0